backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1

# Model Dependencies
spacy==3.8.0
transformers==4.36.2
torch==2.6.0
torchaudio==2.6.0
sentencepiece==0.2.0
safetensors>=0.4.2
scikit-learn==1.4.0
joblib==1.3.2
librosa>=0.10.0
soundfile>=0.12.0

# Utilities
numpy==1.26.3
pandas==2.2.0
python-dotenv==1.0.0
certifi>=2024.2.2
dnspython>=2.6.1

# CORS & Middleware
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
